requirements.txt
boto3>=1.8.6
google-cloud-storage>=1.12.0
jupyter==1.0.0
python-decouple>=3.1
redis>=2.10.6
